Fix 'pushKV ambiguous' error in compilation
- Error occuring because there is no conversion function for size_t to C string in pushKV variatns
- When conversion from size_t to uint64_t doesn't happen automatically on some systems, error is encountered
- Affects at least M1 macOS

+++ src/net.cpp
@@ -2791,7 +2791,7 @@ bool CConnman::RemoveAddedNode(const std::string& strNode)
     return false;
 }
 
-size_t CConnman::GetNodeCount(ConnectionDirection flags) const
+uint64_t CConnman::GetNodeCount(ConnectionDirection flags) const
 {
     LOCK(cs_vNodes);
     if (flags == ConnectionDirection::Both) // Shortcut if we want total
